# tools/repo_orchestrator/openapi.yaml
openapi: 3.1.0
info:
  title: Repo Orchestrator API
  description: >
    Secure read-only access to a selected repository, plus UI/service operations.
    All endpoints require Bearer authentication.
  version: UNRELEASED

servers:
  - url: http://127.0.0.1:9325
    description: Local default

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string

security:
  - bearerAuth: []

paths:
  /status:
    get:
      summary: Service status
      operationId: getStatus
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  version: { type: string }
                  uptime_seconds: { type: number }

  /tree:
    get:
      summary: List directory tree
      operationId: getTree
      parameters:
        - name: path
          in: query
          required: false
          schema: { type: string, default: "." }
        - name: max_depth
          in: query
          required: false
          schema: { type: integer, default: 3, maximum: 6 }
      responses:
        '200':
          description: Files list
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items: { type: string }
                  truncated: { type: boolean }

  /file:
    get:
      summary: Read file content
      operationId: getFile
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string }
        - name: start_line
          in: query
          required: false
          schema: { type: integer, default: 1, minimum: 1 }
        - name: end_line
          in: query
          required: false
          schema: { type: integer, default: 500, minimum: 1 }
      responses:
        '200':
          description: Plaintext content
          content:
            text/plain:
              schema: { type: string }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '413': { description: Too large, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '500': { description: Server error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /search:
    get:
      summary: Search in repository
      operationId: search
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 3, maxLength: 128 }
        - name: ext
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                  truncated: { type: boolean }

  /diff:
    get:
      summary: Git diff summary
      operationId: getDiff
      parameters:
        - name: base
          in: query
          required: false
          schema: { type: string, default: "main" }
        - name: head
          in: query
          required: false
          schema: { type: string, default: "HEAD" }
      responses:
        '200':
          description: Plaintext diff
          content:
            text/plain:
              schema: { type: string }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /ui/status:
    get:
      summary: UI status payload
      operationId: getUiStatus
      responses:
        '200':
          description: OK

  /ui/audit:
    get:
      summary: Tail audit log lines
      operationId: getUiAudit
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 200, minimum: 10, maximum: 500 }
      responses:
        '200':
          description: OK

  /ui/allowlist:
    get:
      summary: Return allowlist items (relative to active repo)
      operationId: getUiAllowlist
      responses:
        '200':
          description: OK

  /ui/repos:
    get:
      summary: List repos under configured root
      operationId: listRepos
      responses:
        '200':
          description: OK

  /ui/repos/active:
    get:
      summary: Get active repo path
      operationId: getActiveRepo
      responses:
        '200':
          description: OK

  /ui/repos/open:
    post:
      summary: Signal opening a repo path (server-agnostic)
      operationId: openRepo
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /ui/repos/select:
    post:
      summary: Select the active repo
      operationId: selectRepo
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /ui/repos/vitaminize:
    post:
      summary: Create supporting files in the target repo (vitaminize)
      operationId: vitaminizeRepo
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /ui/security/events:
    get:
      summary: Get recent security events and panic mode
      operationId: getSecurityEvents
      responses:
        '200':
          description: OK

  /ui/security/resolve:
    post:
      summary: Resolve a security action (e.g. clear panic)
      operationId: resolveSecurity
      parameters:
        - name: action
          in: query
          required: true
          schema: { type: string, enum: ["clear_panic"] }
      responses:
        '200':
          description: OK
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /ui/service/status:
    get:
      summary: Get system service status
      operationId: getServiceStatus
      responses:
        '200':
          description: OK

  /ui/service/restart:
    post:
      summary: Restart system service
      operationId: restartService
      responses:
        '200':
          description: OK
        '500': { description: Server error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /ui/service/stop:
    post:
      summary: Stop system service
      operationId: stopService
      responses:
        '200':
          description: OK
        '500': { description: Server error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }